fix: consolidate default session path tests under pkg/service file cap
Move default-session absolute-path tests into runtime_session_runtime_test.go
so pkg/service stays within the 15-file package limit for make pkg-file-count.
